Added sdhci controller to nexus device tree.
Linked sdhci to a test application so the controller can be tested with the Pi.

Replaced the previous sdhci.c & sdhci.h

Previous sdhci.c and sdhci.h have both the bsp specific code
and generic functions in the file.

The new implementation that is ported, is more modular in nature.